The GalleryVirtuality Project is the brainchild of Bret Colin Sheppard, Designer and Chief Programmer. "I designed this so people in the art
Bret has "disarmed" the Quake 3 engine from it's myriad arsonal and
created a slimmed-down environment for displaying art. No crazed
creatures are there to chase you or shoot at you as you can take your
time, perusing art. Artists can show their work virtually in a familiar
environment to patrons. Teachers and parents can help kids put their art
in a "gallery" inspiring them as future working artists. A walk-through
can be recorded into a YouTube Video to share on social media, can be
used as a movie-making tool, and possibly as an interesting
"presentation" tool. The virtual sky is the limit. That is the idea,
still developing and we are welcoming interested game designers, artists
and others into this project as it comes online in 2015. Those
interested in the code can email bretcolin@gmail.com.